backend/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
python-dotenv==0.21.1